requirements.txt
python-dotenv==1.0.1
langchain==0.2.6
langchain_google_genai==1.0.7
google-generativeai==0.7.1